# swagger.yaml
definitions:
  Activities:
    items:
      $ref: '#/definitions/Activity'
    type: array
    x-go-package: casaattivita/models
  Activity:
    properties:
      data_inserimento:
        description: |-
          Timestamp of insert
          in: time
        format: date-time
        type: string
        x-go-name: DateInsert
      id:
        description: |-
          Id of Message value
          in: int64
        format: int64
        type: integer
        x-go-name: Id
      messaggio:
        description: |-
          Value of Message
          in: int
        type: string
        x-go-name: Value
    type: object
    x-go-name: Message
    x-go-package: casaattivita/models
  CommonError:
    properties:
      message:
        description: |-
          Message of the error
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: casaattivita/controllers
  CommonSuccess:
    properties:
      message:
        description: |-
          Message of the error
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: casaattivita/controllers
  GetActivities:
    properties:
      data:
        $ref: '#/definitions/Activities'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: casaattivita/controllers
  GetActivity:
    properties:
      data:
        $ref: '#/definitions/Activity'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: casaattivita/controllers
  GetIsActive:
    properties:
      active:
        description: |-
          Message of the response
          in: string
        type: boolean
        x-go-name: Active
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: casaattivita/controllers
  GetMessage:
    properties:
      data:
        $ref: '#/definitions/Activity'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: casaattivita/controllers
  GetMessages:
    properties:
      data:
        $ref: '#/definitions/Messages'
      message:
        description: |-
          Message of the response
          in: string
        type: string
        x-go-name: Message
      status:
        description: |-
          Status of the error
          in: int64
        format: int64
        type: integer
        x-go-name: Status
    type: object
    x-go-package: casaattivita/controllers
  Messages:
    items:
      $ref: '#/definitions/Activity'
    type: array
    x-go-package: casaattivita/models
  ReqAddActivity:
    properties:
      valore:
        description: |-
          Value of the Activity
          in: int
        format: int64
        type: integer
        x-go-name: Value
    type: object
    x-go-package: casaattivita/models
info:
  title: 'Activity Api:'
  version: 0.1.3
paths:
  /activities:
    get:
      description: Get Activity list
      operationId: listRain
      responses:
        "200":
          description: GetActivities
          schema:
            $ref: '#/definitions/GetActivities'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      security:
      - apiKey:
        - '[]'
  /activity:
    post:
      description: Create a new Activity value
      operationId: addActivity
      parameters:
      - description: 'name: body'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ReqAddActivity'
        x-go-name: Body
      responses:
        "200":
          description: GetActivity
          schema:
            $ref: '#/definitions/GetActivity'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      security:
      - apiKey:
        - '[]'
  /isactive:
    get:
      description: Get if sensor is online or offline
      operationId: isactive
      responses:
        "200":
          description: GetIsActive
          schema:
            $ref: '#/definitions/GetIsActive'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
  /lasthour:
    get:
      description: Get list of last hour of Activity .... or the last value inserted
      operationId: lastHour
      responses:
        "200":
          description: GetActivities
          schema:
            $ref: '#/definitions/GetActivities'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      security:
      - apiKey:
        - '[]'
  /message:
    post:
      description: Create a new Message value
      operationId: addMessage
      responses:
        "200":
          description: GetMessage
          schema:
            $ref: '#/definitions/GetMessage'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      security:
      - apiKey:
        - '[]'
  /messages:
    get:
      description: Get list of last hour of Messages .... or the last value inserted
      operationId: messages
      responses:
        "200":
          description: GetMessages
          schema:
            $ref: '#/definitions/GetMessages'
        "401":
          description: CommonError
          schema:
            $ref: '#/definitions/CommonError'
      security:
      - apiKey:
        - '[]'
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"